requirements.txt
apache-beam[gcp]
google-cloud-pubsub
google-cloud-bigquery
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
filestack-python
Flask==1.1.2
fpdf==1.7.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
requests>=2.24.0,<3.0.0
trafaret==2.0.2
urllib3==1.25.11
Werkzeug==1.0.1
WTForms==2.3.3
gunicorn==20.1.0
